Move alembic into app for import consistency
