+++ alembic.ini
@@ -1,5 +1,5 @@
 [alembic]
-script_location = alembic
+script_location = app/alembic
 file_template = %%(slug)s_%%(rev)s
 
 [post_write_hooks]
