+++ app/alembic/env.py
@@ -1,6 +1,5 @@
 import os, sys
-sys.path.insert(0, os.path.dirname(__file__))
-sys.path.insert(0, os.path.join(os.path.dirname(__file__), '..'))
+sys.path.insert(0, os.path.join(os.path.dirname(__file__), '../..'))
 
 from logging.config import fileConfig
